fix: update roundToDecimals function to return double and adjust Modbus handling for energy values

+++ source/include/utils.h
@@ -108,9 +108,9 @@ inline float roundToDecimals(float value, uint8_t decimals = 3) {
     float factor = powf(10.0f, decimals);
     return roundf(value * factor) / factor;
 }
-inline float roundToDecimals(double value, uint8_t decimals = 3) {
+inline double roundToDecimals(double value, uint8_t decimals = 3) {
     double factor = pow(10.0, decimals);
-    return (float)(round(value * factor) / factor);
+    return round(value * factor) / factor;
 }
 
 // Device identification
